pyproject.toml
[project]
name = "dazzo-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleak>=1.1.1",
    "python-dotenv>=1.2.1",
    "numpy>=1.26",
    "altair>=5.5.0",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "httpx>=0.28.1",
]
